[PCLI-123]refactor: delete class PeoplesToJson service and add method yours 'execute()' in PeopleService with name 'peoplesToJson'
