+++ routes/people.js
@@ -3,14 +3,12 @@ import { Router } from "express";
 import { PeopleController } from "../controllers/PeopleController.js";
 import { accessLevel } from "../helpers/accessLevel.js";
 import { config } from "dotenv";
-import { PeoplesToJsonService } from "../services/PeoplesToJsonService.js"
 config();
 
 const people = Router();
 people.use(express.static("static"))
 
 const peopleController = new PeopleController()
-const peoplesToJsonService = new PeoplesToJsonService()
 
 people.get("/", (req, res) => 
     peopleController.renderRegister(req, res)
@@ -24,8 +22,8 @@ people.post("/register-people", (req, res) =>
     peopleController.registerPeople(req, res)
 )
 
-people.get("/exportallpeoples", (req, res) => 
-    peoplesToJsonService.execute(req, res)
+people.get("/export-all-peoples", (req, res) => 
+    peopleController.peoplesToJson(req, res)
 )
 
 people.get("/export-peoples", accessLevel.isAdmin, (req, res) =>
