+++ services/PeopleService.js
@@ -1,5 +1,7 @@
 import { Peoples } from "../models/people.js";
 import fs from "fs"
+import { config } from "dotenv";
+config();
 
 class PeopleService {
     
@@ -10,6 +12,21 @@ class PeopleService {
 
     async save(body) {
         await Peoples.create(body);
+    }
+
+    async peoplesToJson(req, res) {     
+        if (req.headers.authorization != process.env.AUTHORIZATION_TOKEN) {
+            return res.status(401).render("unauthorized");
+
+        } else {
+            let result = [];
+            const peoples = await Peoples.findAll();            
+            peoples.forEach(function(people) {
+                result.push(people.dataValues);
+            })
+            res.set("Content-Type", "text/json");
+            return res.status(200).send(result);
+        }
     }
 
     async export() {
